Improve script exit codes
If the scripts --help command was called the default 127 exit code was
returned even if the call was successful. If any other module (like
calling help in the argparse module) throws a SystemExit exception
with the sys.exit() call the exit code was not set properly
because exception was not handled.